refactor: move oauth2 type to core
