+++ packages/better-auth/src/index.ts
@@ -1,6 +1,7 @@
 //#region Re-exports necessaries from core module
 export * from "@better-auth/core/env";
 export * from "@better-auth/core";
+export * from "@better-auth/core/oauth2";
 //#endregion
 export { getCurrentAdapter } from "./context/transaction";
 export * from "./auth";
@@ -14,7 +15,6 @@ export type * from "zod/v4/core";
 //@ts-expect-error: we need to export helper types even when they conflict with better-call types to avoid "The inferred type of 'auth' cannot be named without a reference to..."
 export type * from "./types/helper";
 // export this as we are referencing OAuth2Tokens in the `refresh-token` api as return type
-export type * from "./oauth2/types";
 
 // telemetry exports for CLI and consumers
 export { createTelemetry } from "./telemetry";
