+++ packages/better-auth/src/oauth2/create-authorization-url.ts
@@ -1,4 +1,4 @@
-import type { ProviderOptions } from "./types";
+import type { ProviderOptions } from "@better-auth/core/oauth2";
 import { generateCodeChallenge } from "./utils";
 
 export async function createAuthorizationURL({
