+++ packages/better-auth/src/oauth2/index.ts
@@ -3,5 +3,4 @@ export * from "./validate-authorization-code";
 export * from "./refresh-access-token";
 export * from "./utils";
 export * from "./state";
-export * from "./types";
 export * from "./link-account";
