+++ packages/better-auth/src/oauth2/refresh-access-token.ts
@@ -1,6 +1,5 @@
 import { betterFetch } from "@better-fetch/fetch";
-import type { OAuth2Tokens } from "./types";
-import type { ProviderOptions } from "./types";
+import type { OAuth2Tokens, ProviderOptions } from "@better-auth/core/oauth2";
 import { base64 } from "@better-auth/utils/base64";
 
 export function createRefreshAccessTokenRequest({
