+++ packages/better-auth/src/plugins/generic-oauth/index.ts
@@ -8,10 +8,12 @@ import { BASE_ERROR_CODES } from "../../error/codes";
 import {
 	createAuthorizationURL,
 	validateAuthorizationCode,
-	type OAuth2Tokens,
-	type OAuthProvider,
-	type OAuth2UserInfo,
 } from "../../oauth2";
+import type {
+	OAuth2Tokens,
+	OAuth2UserInfo,
+	OAuthProvider,
+} from "@better-auth/core/oauth2";
 import { handleOAuthUserInfo } from "../../oauth2/link-account";
 import { refreshAccessToken } from "../../oauth2/refresh-access-token";
 import { generateState, parseState } from "../../oauth2/state";
