+++ packages/better-auth/src/social-providers/facebook.ts
@@ -1,5 +1,5 @@
 import { betterFetch } from "@better-fetch/fetch";
-import type { OAuthProvider, ProviderOptions } from "../oauth2";
+import type { OAuthProvider, ProviderOptions } from "@better-auth/core/oauth2";
 import { createAuthorizationURL, validateAuthorizationCode } from "../oauth2";
 import { createRemoteJWKSet, jwtVerify, decodeJwt } from "jose";
 import { refreshAccessToken } from "../oauth2";
