+++ packages/better-auth/src/social-providers/kick.ts
@@ -1,10 +1,6 @@
 import { betterFetch } from "@better-fetch/fetch";
-import {
-	createAuthorizationURL,
-	type OAuthProvider,
-	type ProviderOptions,
-	validateAuthorizationCode,
-} from "../oauth2";
+import { createAuthorizationURL, validateAuthorizationCode } from "../oauth2";
+import type { OAuthProvider, ProviderOptions } from "@better-auth/core/oauth2";
 
 export interface KickProfile {
 	/**
