+++ packages/better-auth/src/social-providers/roblox.ts
@@ -1,5 +1,5 @@
 import { betterFetch } from "@better-fetch/fetch";
-import type { OAuthProvider, ProviderOptions } from "../oauth2";
+import type { OAuthProvider, ProviderOptions } from "@better-auth/core/oauth2";
 import { validateAuthorizationCode, refreshAccessToken } from "../oauth2";
 
 export interface RobloxProfile extends Record<string, any> {
