+++ packages/better-auth/src/social-providers/tiktok.ts
@@ -1,5 +1,5 @@
 import { betterFetch } from "@better-fetch/fetch";
-import type { OAuthProvider, ProviderOptions } from "../oauth2";
+import type { OAuthProvider, ProviderOptions } from "@better-auth/core/oauth2";
 import { refreshAccessToken, validateAuthorizationCode } from "../oauth2";
 
 /**
