+++ packages/better-auth/src/social-providers/twitch.ts
@@ -1,4 +1,4 @@
-import type { OAuthProvider, ProviderOptions } from "../oauth2";
+import type { OAuthProvider, ProviderOptions } from "@better-auth/core/oauth2";
 import { logger } from "@better-auth/core/env";
 import {
 	createAuthorizationURL,
