+++ packages/better-auth/src/social-providers/vk.ts
@@ -1,5 +1,8 @@
 import { betterFetch } from "@better-fetch/fetch";
-import { type OAuthProvider, type ProviderOptions } from "../oauth2";
+import {
+	type OAuthProvider,
+	type ProviderOptions,
+} from "@better-auth/core/oauth2";
 import {
 	createAuthorizationURL,
 	validateAuthorizationCode,
