+++ packages/better-auth/src/social-providers/zoom.ts
@@ -1,10 +1,6 @@
 import { betterFetch } from "@better-fetch/fetch";
-import {
-	type OAuthProvider,
-	type ProviderOptions,
-	generateCodeChallenge,
-	validateAuthorizationCode,
-} from "../oauth2";
+import { generateCodeChallenge, validateAuthorizationCode } from "../oauth2";
+import type { OAuthProvider, ProviderOptions } from "@better-auth/core/oauth2";
 
 export type LoginType =
 	| 0 /** Facebook OAuth */
