+++ packages/core/build.config.ts
@@ -14,5 +14,6 @@ export default defineBuildConfig({
 		"./src/db/adapter/index.ts",
 		"./src/async_hooks/index.ts",
 		"./src/env/index.ts",
+		"./src/oauth2/index.ts",
 	],
 });
